+++ src/dialogs/tododialog.h
@@ -54,6 +54,7 @@ private:
     CalendarItem currentCalendarItem;
     CalendarItem lastCreatedCalendarItem;
     QString _jumpToCalendarItemUid = "";
+    bool _setFocusToDescriptionEdit = false;
     int firstVisibleTodoListRow;
     void setupMainSplitter();
     void storeSettings();