+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 1.2.5
 - in the todo dialog the *save and insert* button will be disabled correctly now
+- did some tab order improvements in the todo dialog
+- the description edit now gets the focus after creating a new task in the 
+  todo dialog
+- fixed the shortcuts in the todo dialog
+    - please visit [QOwnNotes shortcuts](http://www.qownnotes.org/shortcuts/QOwnNotes)
+      for a list of all shortcuts
 
 ## 1.2.4
 - the selected text in a note will be inserted as summary in the todo dialog